requirements.txt
numpy
matplotlib
pandas
pypdf
google-generativeai
nltk
streamlit
rouge
sentence-transformers